Fix encoding of file contents
In Ignition, the "contents" of a File must be encoded as a data URL. RFC
2397 requires this to be urlencoded in a similar manner to a query
string, but with spaces appearing as "%20" instead of "+". To ensure
correct encoding, use the same library that Ignition is using to verify
the URLs to do the encoding.

In Ignition, the "contents" of a systemd unit must *not* be encoded.