Make ServeImage() idempotent
It is guaranteed that this will be called multiple times for identical
image data (writing the URL to the status will trigger the CR to be
reconciled again), so don't generate another one each time.

The controller guarantees that if any of the data changes, the old image
will be removed first. Therefore, it is safe to never update the
contents.