Support initramfs in static server
Serve initramfs images instead of ISO images if the name in the image
mapping does not end with ".iso".

If no name is specified in the image mapping, serve *both* an ISO and an
initramfs. This way we can specify both URLs to ironic via terraform and
the driver will use the preferred one - we won't need to figure out
whether the driver supports ISO or pass configuration to the image
server.