Make Ironic URL optional
If the address for Ironic is not available (e.g. if the metal3 Pod is
not running) then the CBO will start the controller with an empty
IRONIC_BASE_URL. In this case we don't want it to crashloop, but rather
start and update the CRDs to indicate that the images are not available.