Report size of initramfs correctly
The initramfs is created by concatenating the ignition archive on the
end of the base initramfs. Therefore the length depends on the length of
the ignition data.

Previously we were reporting only the length of the base initramfs, with
the result that the ignition was never included when the image was
served.