Set IPv6 ClientID and hostname
For IPv6, we need to use a stable client ID (based on the MAC) so that
the IP address does not change when we move from IPA to the installed
Node. The Node's IP address must match the one in the inspection data
for Certificate Signing Requests to be approved.

We also need to set the hostname to match the one provided by DHCP,
otherwise it will just be "localhost" in the inspection data, and again
Certificate Signing Requests will not be approved.