Add IP_OPTIONS environment variable
The IP_OPTIONS we need to use for the provisioned node (which has to
pull its Ignition from the external network) are not necessarily the
same as the ones used for starting IPA (which has to talk to ironic via
the provisioning network if it exists).

Add an environment variable that can be passed down to the custom deploy
step to set the IP options for the installed system via
coreos-installer.

https://bugzilla.redhat.com/show_bug.cgi?id=2034527